Quibobog/requirements.txt
opencv-python
mediapipe
screeninfo
PyInstaller
pygame-ce   
colorama